+++ NEWS.md
@@ -99,6 +99,6 @@
 
 -   Added **add_overall** argument to **ody_summarise_df**
 
-# odytools 0.6.3 (under development)
+# odytools dev
 
--   Now project-specific functions of a RedCap project are stored in the scripts of the new folder functions/ (no name restrictions).
+-   Now, project-specific functions of a RedCap project are assumed to be in the scripts of the folder functions.