+++ NEWS.md
@@ -103,4 +103,4 @@
 
 -   Now, project-specific functions of a RedCap project are assumed to be in the scripts of the folder functions.
 -   Added **ody_rc_timetravel** to load previous imports.
--   Now, tokens are stored in .Renviron, so you only have to probide it when starting a new project.
+-   Now, tokens are stored in .Renviron, so you only have to provide it when starting a new project or after regenerating your token.