+++ NEWS.md
@@ -106,3 +106,4 @@
 -   Now, tokens are stored in .Renviron, so you only have to provide it when starting a new project or after regenerating your token.
 -   Added **ody_extract_km** to get the median and confidence intervals of survfit objects.
 -   Added **ody_proj_init** to set the basicc structure of a non-RedCap project.
+-   Now, **ody_add_version** looks for already rendered outputs in the specified path instead than only in the current wd (the default path).