+++ NEWS.md
@@ -107,3 +107,4 @@
 -   Added **ody_extract_km** to get the median and confidence intervals of survfit objects.
 -   Added **ody_proj_init** to set the basicc structure of a non-RedCap project.
 -   Now, **ody_add_version** looks for already rendered outputs in the specified path instead than only in the current wd (the default path).
+-   Added **Create Lockfile** addin. It creates and saves a lockfile of the current project, so if any thing crashes after an update, the project can we isolated creating its own library according to the lockfile.